justcal-desktop/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.9.2
keyring==25.4.1
sentence-transformers==3.1.1
numpy==1.26.4
PyMuPDF==1.24.10
python-docx==1.1.2
